Always reuse available layer to avoid pauses
If restoring from PiP to a view using the same player as the one playing in PiP, we must always avoid instantiating a second layer to attach it to the player.

If a player is namely attached to two layers, one in PiP, stopping PiP will namely pause playback. This does not happen if we have the same layer (the PiP overlay will simply be merged back).